requirements.txt
anyio==4.3.0
base58==2.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
construct==2.10.68
construct-typing==0.5.6
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
jsonalias==0.1.1
numpy==1.26.4
pandas==2.2.1
portalocker==2.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
solana==0.33.0
solders==0.21.0
tabulate==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
websockets==11.0.3
